add: timeline editor components seperated, TimelineEditor is 600 lines now
